+++ DennMutation.h
@@ -21,7 +21,7 @@ namespace Denn
 		//Mutation
 		Mutation(const DennAlgorithm& algorithm);
 		//operation
-		virtual void operator()(const Population& population,int id_target,Individual& output)= 0; 
+		virtual void operator()(const Population& population,size_t id_target,Individual& output)= 0; 
 		//required sort (default false)
 		virtual bool required_sort() const { return false; }
